Create StoriesGrid component and use in search stories and overview
